requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-cohere
chromadb
google-generativeai
pypdf
python-multipart
pytest
requests
python-docx
beautifulsoup4
pywin32